Move keying of blob names into separate functions.
This allows us to configure whether the instance name should be taken
into account when storing blobs. This is useful for the CAS, where we
can choose to ignore the instance name. For the AC it is important that
the instance name is preserved, as different instances can generate
different output.