Add a container image for a frontend process.
The frontend process will be responsible for dealing with RPCs from the
Bazel tool. By talking to S3 directly, it'll be capable of dealing with
all of the CAS, AC and ByteStream RPCs. As all of this is stateless, it
can easily run with more replicas.

We'll need to extend this to send all Execution and Watch RPCs to some
kind of scheduler process. This process can only run once, but can have
a pool of workers behind it.