Let workers call FindMissing() to prevent unnecessary uploads.
Add a batch writing blob adapter that piles up files that need to be
written. When we have a bunch of them, call FindMissing() and see what
actually needs to be uploaded.

Unfortunately, this seems to destroy performance instead of improving
it. Place it in a separate branch to ensure it won't get lost. We should
revisit it when we've figured out how to optimize FindMissing() on S3.